# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Catch API
  version: 0.1.0
  description: Backend contracts for Catch modular monolith.
servers:
  - url: /api/v1
...
paths:
...
  /articles/{articleId}/comments:
    get:
      tags: [Comments]
      summary: List article comments
      operationId: listArticleComments
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          description: Article comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentList"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Comments]
      summary: Create article comment
      operationId: createArticleComment
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /comments/{commentId}:
    get:
      tags: [Comments]
      summary: Get comment by id
      operationId: getComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Comments]
      summary: Edit own comment during edit window
      operationId: updateComment
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /chat/conversations:
    get:
      tags: [Chat]
      summary: List current user chat conversations
      operationId: listChatConversations
      security:
        - cookieSession: []
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Chat conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatConversationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Chat]
      summary: Start or return direct chat conversation
      operationId: startChatConversation
      security:
        - cookieSession: []
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartChatConversationRequest"
      responses:
        "201":
          description: Chat conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatConversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /chat/conversations/{conversationID}/messages:
    get:
      tags: [Chat]
      summary: List chat messages
      operationId: listChatMessages
      security:
        - cookieSession: []
      parameters:
        - name: conversationID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Chat]
      summary: Send chat message
      operationId: sendChatMessage
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - name: conversationID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageRequest"
      responses:
        "201":
          description: Chat message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /chat/conversations/{conversationID}/messages/stream:
    get:
      tags: [Chat]
      summary: Stream chat messages
      operationId: streamChatMessages
      security:
        - cookieSession: []
      parameters:
        - name: conversationID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: after_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Server-sent events with ready, ping and message events
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chat/conversations/{conversationID}/read:
    post:
      tags: [Chat]
      summary: Mark chat conversation as read
      operationId: markChatConversationRead
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - name: conversationID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Conversation marked read
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: catch_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    ArticleId:
      name: articleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated user has no permission for this action
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Service is temporarily unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    CreateCommentRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        parent_id:
          type: string
          format: uuid
        body:
          type: string
          minLength: 1
          maxLength: 4000
    UpdateCommentRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 4000
    Comment:
      type: object
      additionalProperties: false
      required: [id, article_id, author_id, body, status, reactions_up, reactions_down, reaction_score, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        article_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
        body:
          type: string
        status:
          type: string
          enum: [active, deleted]
        reactions_up:
          type: integer
          minimum: 0
        reactions_down:
          type: integer
          minimum: 0
        reaction_score:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        edited_at:
          type: string
          format: date-time
    CommentList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
...
    StartChatConversationRequest:
      type: object
      additionalProperties: false
      required: [recipient_id]
      properties:
        recipient_id:
          type: string
          format: uuid
    SendChatMessageRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 4000
    ChatConversation:
      type: object
      additionalProperties: false
      required: [id, member_ids, unread_count, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        member_ids:
          type: array
          items:
            type: string
            format: uuid
        unread_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChatConversationList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChatConversation"
    ChatMessage:
      type: object
      additionalProperties: false
      required: [id, conversation_id, sender_id, body, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        body:
          type: string
        status:
          type: string
          enum: [sent, read]
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
    ChatMessageList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [type, title, status, code, message]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - not_implemented
            - service_unavailable
            - internal_error
        message:
          type: string
        trace_id:
          type: string
        details:
          type: object
          additionalProperties: true